Add support for TextureViews as per issue #57
As per owner's suggestion:
Firstly get all children recursively;
Then check if each child is an instance of TextureView with instanceof;
Make sure child knows it's not opaque and get its bitmap into a buffer;
Draw said bitmap into the current canvas.

This assumes that these TextureView elements are ordered in the tree.